+++ lib/gg_router.dart
@@ -4,15 +4,12 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-export 'src/gg_route_tree_node.dart';
-
-export 'src/gg_router_delegate.dart';
-export 'src/gg_route_information_parser.dart';
-
-export 'src/gg_router.dart';
-export 'src/gg_popover_route.dart';
-export 'src/gg_route_params.dart';
-
-export 'src/gg_route_change_builder.dart';
-
+export 'src/gg_animations.dart';
 export 'src/gg_navigation_page.dart';
+export 'src/gg_popover_route.dart';
+export 'src/gg_route_change_builder.dart';
+export 'src/gg_route_information_parser.dart';
+export 'src/gg_route_params.dart';
+export 'src/gg_route_tree_node.dart';
+export 'src/gg_router.dart';
+export 'src/gg_router_delegate.dart';