+++ test/gg_route_tree_node_test.dart
@@ -32,6 +32,7 @@ main() {
 
   dispose() {
     root.dispose();
+    expect(root.isDisposed, isTrue);
   }
 
   group('Param', () {